+++ RingBuffer.h
@@ -6,13 +6,13 @@
 class RingBuffer{
 public:
     int capacity_;
-    double* buffer;
+    float* buffer;
     int head;
     int tail;
     int count_;
 
     RingBuffer(const int&  capacity){
-        buffer = new double[capacity];
+        buffer = new float[capacity];
         capacity_ = capacity;
         for (int i = 0; i < capacity_; i++){
             buffer[i] = 0;
@@ -29,17 +29,17 @@ public:
     bool isEmpty() { return count_ == 0; }
     bool isFull() { return count_ == capacity_; }
 
-    void push(const double& value){
+    void push(const float& value){
         if(isFull()){ head = (head+1) % capacity_;}
         else{count_++;}
         buffer[tail] = value;
         tail = (tail+1)% capacity_;
     }
 
-    double pop(){
+    float pop(){
         if(isEmpty()){ std::cout << "Buffer is empty" << std::endl; return -1;}
         else{ count_--;}
-        double temp = buffer[head];
+        float temp = buffer[head];
         buffer[head] = 0;
         head = (head+1) % capacity_;
         return temp;
